refactor:  refactor cli/args and implement draft modules/cjsToEsm
