+++ src/modules/asd.js
@@ -1,0 +1,4 @@
+import * as myModule from './cjsToEsm.mjs';
+
+console.log(myModule.unknownObject);
+myModule.createMyServer.listen(8000);
