+++ src/modules/files/a.json
@@ -1,4 +1,4 @@
-{
+ {
     "a": 1,
     "b": 2,
     "c": 3
